B1

+++ go_fast_tutorial/tutorial_6/method1.go
@@ -3,16 +3,17 @@ package main
 import "fmt"
 
 type icEngine struct{
-	kml uint8
-	litres uint8
+	kml uint16
+	litres uint16
 }
 
-func (e icEngine) kmsleft() uint8{
+func (e icEngine) kmsleft() uint16{
 	return e.litres*e.kml
 }
 
 func main(){
 	var myICE icEngine = icEngine{25, 15}
+	// 25*15 gives integer overflow error due to uint8 taking only upto 0-255 value. [performs wrong calculation]
 	fmt.Printf("Total kms left in tank: %v", myICE.kmsleft()) 
 }
